+++ packages/core/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe core SDK package will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Added `certAccount` option to base profile options to support client certificate authentication using certificates stored in system keystores on macOS and Windows platforms. [#2325](https://github.com/zowe/zowe-cli/issues/2325)
+
 ## `8.31.0`
 
 - Enhancement: Added a `requestQueue` queue object to the `ZosmfRestClient` class. [#2543](https://github.com/zowe/zowe-cli/pull/2543)